pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "papermentor-os"
version = "0.1.0"
description = "PaperMentor OS backend skeleton and MVP review pipeline"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "pydantic>=2.10,<3.0",
  "python-docx>=1.1.2,<2.0",
  "pyyaml>=6.0,<7.0",
  "uvicorn>=0.34,<1.0",
  "langgraph>=0.2,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/papermentor_os"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
